backend/requirements.txt
altgraph==0.17.2
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
future==0.18.2
macholib==1.16.3
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.0
gunicorn
